backend/requirements.txt
pydantic
fastapi
uvicorn
sqlalchemy
psycopg2-binary
asyncpg
python-dotenv
openai
logtail-python
starlette
pyjwt
httpx
fastapi-jwt-auth
openpyxl
pydantic[email]
bcrypt==3.2.2
PyMuPDF
passlib==1.7.4
PyPDF2
aiofiles
python-multipart
azure-storage-blob
aiohttp
langchain
langchain-openai
langchain_core